requirements.txt
Django>=2.0,<2.2
dnspython3==1.12.0
psycopg2-binary>=2.7.6
django-password-reset==2.0
djangorestframework==3.9.1
